+++ dwsys/DoublyLinkedList.h
@@ -2,7 +2,7 @@
 #define _DoublyLinkedList_h_
 /* DoublyLinkedList.h
  *
- * Copyright (C) 2011, 2015 David Weenink, 2015 Paul Boersma
+ * Copyright (C) 2011 - 2017 David Weenink, 2015 Paul Boersma
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -32,7 +32,7 @@ Thing_define (DoublyLinkedNode, Daata) {
 };
 
 Thing_define (DoublyLinkedList, Thing) {
-	long numberOfNodes;
+	integer numberOfNodes;
 	DoublyLinkedNode front, back;
 
 	void v_destroy () noexcept
