+++ dwtools/DTW_and_TextGrid.cpp
@@ -37,7 +37,7 @@ autoTextTier DTW_and_TextTier_to_TextTier (DTW me, TextTier thee, double precisi
 			autoTextTier him = Data_copy (thee);
 			his xmin = my xmin;
 			his xmax = my xmax;
-			for (long i = 1; i <= his points.size; i ++) {
+			for (integer i = 1; i <= his points.size; i ++) {
 				TextPoint textpoint = his points.at [i];
 				double time = DTW_getXTimeFromYTime (me, textpoint -> number);
 				textpoint -> number = time;
@@ -47,7 +47,7 @@ autoTextTier DTW_and_TextTier_to_TextTier (DTW me, TextTier thee, double precisi
 			autoTextTier him = Data_copy (thee);
 			his xmin = my ymin;
 			his xmax = my ymax;
-			for (long i = 1; i <= his points.size; i ++) {
+			for (integer i = 1; i <= his points.size; i ++) {
 				TextPoint textpoint = his points.at [i];
 				double time = DTW_getYTimeFromXTime (me, textpoint -> number);
 				textpoint -> number = time;
@@ -67,7 +67,7 @@ autoIntervalTier DTW_and_IntervalTier_to_IntervalTier (DTW me, IntervalTier thee
 			autoIntervalTier him = Data_copy (thee);
 			his xmin = my xmin;
 			his xmax = my xmax;
-			for (long i = 1; i <= his intervals.size; i ++) {
+			for (integer i = 1; i <= his intervals.size; i ++) {
 				TextInterval textinterval = his intervals.at [i];
 				double xmin = DTW_getXTimeFromYTime (me, textinterval -> xmin);
 				textinterval -> xmin = xmin;
@@ -79,7 +79,7 @@ autoIntervalTier DTW_and_IntervalTier_to_IntervalTier (DTW me, IntervalTier thee
 			autoIntervalTier him = Data_copy (thee);
 			his xmin = my ymin;
 			his xmax = my ymax;
-			for (long i = 1; i <= his intervals.size; i ++) {
+			for (integer i = 1; i <= his intervals.size; i ++) {
 				TextInterval textinterval = his intervals.at [i];
 				double xmin = DTW_getYTimeFromXTime (me, textinterval -> xmin);
 				textinterval -> xmin = xmin;
@@ -110,7 +110,7 @@ autoTextGrid DTW_and_TextGrid_to_TextGrid (DTW me, TextGrid thee, double precisi
 
 		autoTextGrid him = TextGrid_createWithoutTiers (tmin, tmax);
 
-		for (long i = 1; i <= thy tiers->size; i ++) {
+		for (integer i = 1; i <= thy tiers->size; i ++) {
 			Function anyTier = thy tiers->at [i];
 
 			if (anyTier -> classInfo == classIntervalTier) {
@@ -131,11 +131,11 @@ autoTextGrid DTW_and_TextGrid_to_TextGrid (DTW me, TextGrid thee, double precisi
 
 autoTable DTW_and_IntervalTier_to_Table (DTW me, IntervalTier thee, double precision) {
 	try {
-		long numberOfIntervals = thy intervals.size;
+		integer numberOfIntervals = thy intervals.size;
 		autoTable him = Table_createWithColumnNames (numberOfIntervals, U"tmin tmax label dist");
 		if (fabs (my ymin - thy xmin) <= precision && fabs (my ymax - thy xmax) <= precision) { // map from Y to X
-			long pathIndex = 1;
-			for (long i = 1; i <= numberOfIntervals; i ++) {
+			integer pathIndex = 1;
+			for (integer i = 1; i <= numberOfIntervals; i ++) {
 				TextInterval textinterval = thy intervals.at [i];
 				double xmin = DTW_getXTimeFromYTime (me, textinterval -> xmin);
 				double xmax = DTW_getXTimeFromYTime (me, textinterval -> xmax);
@@ -152,8 +152,8 @@ autoTable DTW_and_IntervalTier_to_Table (DTW me, IntervalTier thee, double preci
 				Table_setNumericValue (him.get(), i, 4, sumOfDistances / numberOfFrames);
 			}
 		} else if (fabs (my xmin - thy xmin) <= precision && fabs (my xmax - thy xmax) <= precision) {  // map from X to Y
-			long pathIndex = 1;
-			for (long i = 1; i <= numberOfIntervals; i ++) {
+			integer pathIndex = 1;
+			for (integer i = 1; i <= numberOfIntervals; i ++) {
 				TextInterval textinterval = thy intervals.at [i];
 				double ymin = DTW_getYTimeFromXTime (me, textinterval -> xmin);
 				double ymax = DTW_getYTimeFromXTime (me, textinterval -> xmax);
@@ -188,7 +188,7 @@ autoTextTier DTW_and_TextTier_to_TextTier_old (DTW me, TextTier thee) {
 		his xmin = my ymin;
 		his xmax = my ymax;
 
-		for (long i = 1; i <= his points.size; i ++) {
+		for (integer i = 1; i <= his points.size; i ++) {
 			TextPoint textpoint = his points.at [i];
 			double time = DTW_getYTimeFromXTime (me, textpoint -> number);
 			textpoint -> number = time;
@@ -209,7 +209,7 @@ autoIntervalTier DTW_and_IntervalTier_to_IntervalTier_old (DTW me, IntervalTier 
 		his xmin = my ymin;
 		his xmax = my ymax;
 
-		for (long i = 1; i <= his intervals.size; i ++) {
+		for (integer i = 1; i <= his intervals.size; i ++) {
 			TextInterval textinterval = his intervals.at [i];
 			double xmin = DTW_getYTimeFromXTime (me, textinterval -> xmin);
 			textinterval -> xmin = xmin;
@@ -233,7 +233,7 @@ autoTextGrid DTW_and_TextGrid_to_TextGrid_old (DTW me, TextGrid thee) {
 		his xmax = my ymax;
 		his tiers = FunctionList_create ();
 
-		for (long i = 1; i <= thy tiers->size; i ++) {
+		for (integer i = 1; i <= thy tiers->size; i ++) {
 			Daata anyTier = thy tiers->at [i];
 
 			if (anyTier -> classInfo == classIntervalTier) {
